# Optimize Project.visible_by_course