Refactor to retrieve all collaborations and pass through to can_read.
Discovering the joys of prefetch_related for GenericForeignKeys (!)
also led to some good performance increases. 

Testing in Music Humanities, which has a few hundred projects and lots of selections, and through a browser request to pick up the django-debug-toolbar metrics.

Initial Benchmark (from local dev, debug environment)
CPU: 25902ms
SQL: 2033 q in 2488ms

Refactor
CPU: 14671ms
SQL: 1284 q in 1355ms